Begin trying to implement gapless playback using singletrakc POC
Since the position is reported using an event, one can compare the percentage of progress to try and start the next track before the "gap" in vlc. The implementation is not the best, but is once again a proof of concept. Even if this is reverted to be able to merge the player, the playback reporting option can make it so that this can replace the old player and a seekbar can be added.